feat: sprite animation and pointer ownership improvement

+++ Project1/Sprite.h
@@ -5,10 +5,10 @@
 class Sprite
 {
     public:
-    int width, height;
+    uint8_t width, height;
     std::vector<Pixel> data;
 
-    Sprite(int w, int h, int posX, int posY) : width(w), height(h), data(w* h) {}
+    Sprite(uint8_t w, uint8_t h, int posX, int posY) : width(w), height(h), data(w* h) {}
 
     void setPixel(int x, int y, Pixel p);
